pyproject.toml
[tool.poetry]
name = "metamorphic-api"
version = "0.1.0"
description = "metamorphic api"
authors = ["blauyourmind <michaelblau21@gmail.com>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
web3 = "^5.28.0"
fastapi = {extras = ["all"], version = "^0.75.0"}
python-dotenv = "^0.19.2"
streamlit = "^1.7.0"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
